schema: Decouple context functions
